+++ aio_pika/connection.py
@@ -2,17 +2,17 @@ import asyncio
 import logging
 from contextlib import suppress
 from functools import wraps, partial
-from typing import Callable, Any, Generator
+from typing import Callable
 
-import pika.channel
-from pika import ConnectionParameters
-from pika.credentials import ExternalCredentials, PlainCredentials
-from pika.spec import REPLY_SUCCESS
+from .pika.channel import Channel as PikaChannel
+from .pika import ConnectionParameters
+from .pika.credentials import ExternalCredentials, PlainCredentials
+from .pika.spec import REPLY_SUCCESS
 from yarl import URL
 from . import exceptions
 from .channel import Channel
 from .common import FutureStore
-from .tools import create_future
+from .tools import shield
 from .adapter import AsyncioConnection
 
 try:
@@ -48,14 +48,18 @@ class Connection:
 
     CHANNEL_CLASS = Channel
 
-    def __init__(self, host: str = 'localhost', port: int=5672, login: str = 'guest',
-                 password: str = 'guest', virtual_host: str = '/',
-                 ssl: bool=False, *, loop=None, **kwargs):
+    def __init__(self, host: str = 'localhost', port: int=5672,
+                 login: str = 'guest', password: str = 'guest',
+                 virtual_host: str = '/', ssl: bool=False, *,
+                 loop=None, **kwargs):
 
         self.loop = loop if loop else asyncio.get_event_loop()
         self.future_store = FutureStore(loop=self.loop)
 
-        self.__credentials = PlainCredentials(login, password) if login else ExternalCredentials()
+        self.__credentials = (
+            PlainCredentials(login, password)
+            if login else ExternalCredentials()
+        )
 
         self.__connection_parameters = ConnectionParameters(
             host=host,
@@ -73,8 +77,10 @@ class Connection:
 
     def __str__(self):
         return 'amqp://{credentials}{host}:{port}/{vhost}'.format(
-            credentials="{0.username}:********@".format(self.__credentials) if isinstance(
-                self.__credentials, PlainCredentials) else '',
+            credentials=(
+                "{0.username}:********@".format(self.__credentials)
+                if isinstance(self.__credentials, PlainCredentials) else ''
+            ),
             host=self.__connection_parameters.host,
             port=self.__connection_parameters.port,
             vhost=self.__connection_parameters.virtual_host,
@@ -131,8 +137,7 @@ class Connection:
             self.__closing = self.future_store.create_future()
 
     @property
-    @asyncio.coroutine
-    def closing(self):
+    async def closing(self):
         """ Return coroutine which will be finished after connection close.
 
         Example:
@@ -154,19 +159,28 @@ class Connection:
                 await connection.closing
 
         """
-        return (yield from self._closing)
+        return await self._closing
 
     def _on_channel_open(self, channel: Channel):
         self._channels[channel.number] = channel
 
-    def _channel_cleanup(self, channel: pika.channel.Channel):
+    def _channel_cleanup(self, channel: PikaChannel):
         ch = self._channels.pop(channel.channel_number)     # type: Channel
         ch._futures.reject_all(exceptions.ChannelClosed)
 
-    def _on_connection_refused(self, future: asyncio.Future, connection: AsyncioConnection, message: str):
-        self._on_connection_lost(future, connection, code=500, reason=ConnectionRefusedError(message))
+    def _on_connection_open(self, future: asyncio.Future,
+                            connection: AsyncioConnection):
+        log.debug("Connection ready: %r", self)
+        if not future.done():
+            future.set_result(connection)
 
-    def _on_connection_lost(self, future: asyncio.Future, connection: AsyncioConnection, code, reason):
+    def _on_connection_refused(self, future: asyncio.Future,
+                               connection: AsyncioConnection, message: str):
+        self._on_connection_lost(future, connection, code=500,
+                                 reason=ConnectionRefusedError(message))
+
+    def _on_connection_lost(self, future: asyncio.Future,
+                            connection: AsyncioConnection, code, reason):
         if self.__closing and self.__closing.done():
             return
 
@@ -185,33 +199,33 @@ class Connection:
 
         future.set_exception(exc)
 
-    def _on_channel_cancel(self, channel: pika.channel.Channel):
+    def _on_channel_cancel(self, channel: PikaChannel):
         ch = self._channels.pop(channel.channel_number)     # type: Channel
         ch._futures.reject_all(exceptions.ChannelClosed)
 
-    @asyncio.coroutine
-    def connect(self) -> AsyncioConnection:
-        """ Connect to AMQP server. This method should be called after :func:`aio_pika.connection.Connection.__init__`
+    async def connect(self) -> AsyncioConnection:
+        """ Connect to AMQP server. This method should be called after
+        :func:`aio_pika.connection.Connection.__init__`
 
         .. note::
-            This method is called by :func:`connect`. You shouldn't call it explicitly.
+            This method is called by :func:`connect`.
+            You shouldn't call it explicitly.
 
         """
-
         if self.__closing and self.__closing.done():
             raise RuntimeError("Invalid connection state")
 
-        with (yield from self.__write_lock):
+        async with self.__write_lock:
             self._connection = None
 
             log.debug("Creating a new AMQP connection: %s", self)
 
-            f = create_future(loop=self.loop)
+            f = self.loop.create_future()
 
             connection = AsyncioConnection(
                 parameters=self.__connection_parameters,
                 loop=self.loop,
-                on_open_callback=f.set_result,
+                on_open_callback=partial(self._on_connection_open, f),
                 on_close_callback=partial(self._on_connection_lost, f),
                 on_open_error_callback=partial(self._on_connection_refused, f),
             )
@@ -219,16 +233,15 @@ class Connection:
             connection.channel_cleanup_callback = self._channel_cleanup
             connection.channel_cancel_callback = self._on_channel_cancel
 
-            result = yield from f
-
-            log.debug("Connection ready: %r", self)
+            result = await f
 
             self._connection = connection
             return result
 
     @_ensure_connection
-    def channel(self, channel_number: int=None, publisher_confirms: bool=True,
-                on_return_raises=False) -> Generator[Any, None, Channel]:
+    def channel(self, channel_number: int=None,
+                publisher_confirms: bool=True,
+                on_return_raises=False) -> Channel:
         """ Coroutine which returns new instance of :class:`Channel`.
 
         Example:
@@ -250,7 +263,9 @@ class Connection:
                 await channel42.close()
 
                 # For working with transactions
-                channel_no_confirms = connection.channel(publisher_confirms=True)
+                channel_no_confirms = connection.channel(
+                    publisher_confirms=True
+                )
                 await channel_no_confirms.close()
 
         Also available as an asynchronous context manager:
@@ -271,9 +286,10 @@ class Connection:
 
         :param channel_number: specify the channel number explicit
         :param publisher_confirms:
-            if `True` the :func:`aio_pika.Exchange.publish` method will be return
-            :class:`bool` after publish is complete. Otherwise the
-            :func:`aio_pika.Exchange.publish` method will be return :class:`None`
+            if `True` the :func:`aio_pika.Exchange.publish` method will be
+            return :class:`bool` after publish is complete. Otherwise the
+            :func:`aio_pika.Exchange.publish` method will be return
+            :class:`None`
         :param on_return_raises:
             raise an :class:`aio_pika.exceptions.UnroutableError`
             when mandatory message will be returned
@@ -286,18 +302,20 @@ class Connection:
                                      on_return_raises=on_return_raises)
 
         log.debug("Channel created: %r", channel)
-
         return channel
 
     def close(self) -> asyncio.Task:
         """ Close AMQP connection """
         log.debug("Closing AMQP connection")
 
-        @asyncio.coroutine
-        def inner():
+        if self.is_closed:
+            return self._closing
+
+        @shield
+        async def inner():
             if self._connection:
                 self._connection.close()
-            yield from self.closing
+            await self.closing
 
         return self.loop.create_task(inner())
 
@@ -306,24 +324,22 @@ class Connection:
             if not self.is_closed:
                 self.close()
 
-    @asyncio.coroutine
-    def __aenter__(self) -> 'Connection':
+    async def __aenter__(self) -> 'Connection':
         return self
 
-    @asyncio.coroutine
-    def __aexit__(self, exc_type, exc_val, exc_tb):
+    async def __aexit__(self, exc_type, exc_val, exc_tb):
         for channel in tuple(self._channels.values()):
-            yield from channel.close()
+            await channel.close()
 
-        yield from self.close()
+        await self.close()
 
 
-@asyncio.coroutine
-def connect(url: str=None, *, host: str='localhost',
-            port: int=5672, login: str='guest',
-            password: str='guest', virtualhost: str='/',
-            ssl: bool=False, loop=None, ssl_options: dict=None,
-            connection_class=Connection, **kwargs) -> Generator[Any, None, Connection]:
+async def connect(url: str=None, *, host: str='localhost', port: int=5672,
+                  login: str='guest', password: str='guest',
+                  virtualhost: str='/', ssl: bool=False, loop=None,
+                  ssl_options: dict=None, connection_class=Connection,
+                  **kwargs) -> Connection:
+
     """ Make connection to the broker.
 
     Example:
@@ -355,33 +371,36 @@ def connect(url: str=None, *, host: str='localhost',
             * ssl_version
 
         For an information on what the ssl_options can be set to reference the
-        `official Python documentation`_.
-
-        .. _official Python documentation: http://docs.python.org/3/library/ssl.html
+        `official Python documentation`_ .
 
     URL string might be contain ssl parameters e.g.
-    `amqps://user:password@10.0.0.1//?ca_certs=ca.pem&certfile=cert.pem&keyfile=key.pem`
+    `amqps://user:pass@host//?ca_certs=ca.pem&certfile=crt.pem&keyfile=key.pem`
 
-    :param url: `RFC3986`_ formatted broker address. When :class:`None` \
-                will be used keyword arguments.
+    :param url:
+        RFC3986_ formatted broker address. When :class:`None`
+        will be used keyword arguments.
     :param host: hostname of the broker
     :param port: broker port 5672 by default
-    :param login: username string. `'guest'` by default. Provide empty string \
-                  for pika.credentials.ExternalCredentials usage.
+    :param login:
+        username string. `'guest'` by default. Provide empty string
+        for pika.credentials.ExternalCredentials usage.
     :param password: password string. `'guest'` by default.
     :param virtualhost: virtualhost parameter. `'/'` by default
-    :param ssl: use SSL for connection. Should be used with addition kwargs. \
-                See `pika documentation`_ for more info.
+    :param ssl:
+        use SSL for connection. Should be used with addition kwargs.
+        See `pika documentation`_ for more info.
     :param ssl_options: A dict of values for the SSL connection.
-    :param loop: Event loop (:func:`asyncio.get_event_loop()` \
-                 when :class:`None`)
+    :param loop:
+        Event loop (:func:`asyncio.get_event_loop()` when :class:`None`)
     :param connection_class: Factory of a new connection
-    :param kwargs: addition parameters which will be passed to \
-                   the pika connection.
+    :param kwargs:
+        addition parameters which will be passed to the pika connection.
     :return: :class:`aio_pika.connection.Connection`
 
-    .. _RFC3986: https://tools.ietf.org/html/rfc3986
+    .. _RFC3986: https://goo.gl/MzgYAs
     .. _pika documentation: https://goo.gl/TdVuZ9
+    .. _official Python documentation: https://goo.gl/pty9xA
+
 
     """
 
@@ -417,7 +436,7 @@ def connect(url: str=None, *, host: str='localhost',
         ssl_options=ssl_options, **kwargs
     )
 
-    yield from connection.connect()
+    await connection.connect()
     return connection