+++ docs/source/examples/master.py
@@ -8,22 +8,21 @@ async def main():
         "amqp://guest:guest@127.0.0.1/"
     )
 
-    # Creating channel
-    channel = await connection.channel()
+    async with connection:
+        # Creating channel
+        channel = await connection.channel()
 
-    master = Master(channel)
+        master = Master(channel)
 
-    # Creates tasks by proxy object
-    for task_id in range(1000):
-        await master.proxy.my_task_name(task_id=task_id)
+        # Creates tasks by proxy object
+        for task_id in range(1000):
+            await master.proxy.my_task_name(task_id=task_id)
 
-    # Or using create_task method
-    for task_id in range(1000):
-        await master.create_task(
-            'my_task_name', kwargs=dict(task_id=task_id)
-        )
-
-    await connection.close()
+        # Or using create_task method
+        for task_id in range(1000):
+            await master.create_task(
+                'my_task_name', kwargs=dict(task_id=task_id)
+            )
 
 
 if __name__ == "__main__":