+++ tests/test_amqp_without_publisher_confirms_robust.py
@@ -7,6 +7,9 @@ from tests import AMQP_URL
 from tests.test_amqp_without_publisher_confirms import TestCase as AMQPTestCase
 
 
+pytestmark = pytest.mark.asyncio
+
+
 class TestCase(AMQPTestCase):
     @asyncio.coroutine
     def create_connection(self, cleanup=True):
@@ -18,11 +21,14 @@ class TestCase(AMQPTestCase):
         return client
 
     @asyncio.coroutine
-    def create_channel(self, connection=None, cleanup=False, publisher_confirms=False, **kwargs):
+    def create_channel(self, connection=None, cleanup=False,
+                       publisher_confirms=False, **kwargs):
         if connection is None:
             connection = yield from self.create_connection()
 
-        channel = yield from connection.channel(publisher_confirms=publisher_confirms, **kwargs)
+        channel = yield from connection.channel(
+            publisher_confirms=publisher_confirms, **kwargs
+        )
 
         if cleanup:
             self.addCleanup(channel.close)
@@ -30,6 +36,7 @@ class TestCase(AMQPTestCase):
         return channel
 
     @pytest.mark.asyncio
+    @asyncio.coroutine
     def test_set_qos(self):
         channel = yield from self.create_channel()
         yield from channel.set_qos(prefetch_count=1)